+++ Models/Identifiers/AbilityIdentifier.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using DiskCardGame;
 
@@ -17,7 +18,13 @@ namespace APIPlugin
 			ids.Add(this);
 		}
 
+		[Obsolete("Replaced by GetID")]
 		public static AbilityIdentifier GetAbilityIdentifier(string guid, string name)
+		{
+			return GetID(guid, name);
+		}
+
+		public static AbilityIdentifier GetID(string guid, string name)
 		{
 			return ids.Exists(x => x.guid == guid && x.name == name)
 				? ids.Find(x => x.guid == guid && x.name == name)