+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## 1.13.0
+- Added support for custom card backgrounds, dialogs, encounters and talking cards
+- Fixes to abilities loading and stackable custom abilities
+
 ## v1.12.1
 - Bugfix so CustomCard doesn't wipe ability information.